+++ gulp/tasks/build.js
@@ -7,7 +7,7 @@ gulp.task('build', 'Builds the code along with running quality checks (tests, hi
     'i18n',
     'code-coverage',
     ['sass', 'html-minify', 'browserify-production', 'copy-static-assets'],
-    //'static-rewrite',
+    'assets-rewrite',
     'complexity',
     done
   );
@@ -17,7 +17,7 @@ gulp.task('build-quick', 'Builds the code without running quality checks', funct
   runSequence(
     'i18n',
     ['sass', 'html-minify', 'browserify', 'copy-static-assets'],
-    //'static-rewrite',
+    'assets-rewrite',
     done
   );
 });