+++ web/app/components/core/application.component.jsx
@@ -2,6 +2,9 @@ var React = require('react/addons');
 var Header = require('./header.component.jsx');
 var RouteHandler = require('react-router').RouteHandler;
 
+//TODO: remove: for testing assets rewrit eprocess
+var removeMe = '/app/components/core/remove-me.svg#icon-small';
+
 var Application = React.createClass({
   render: function applicationComponentRender() {
     return (